update controller & security

+++ src/main/java/com/shoes_store/lenhan/controller/brandController.java
@@ -1,6 +1,7 @@
 package com.shoes_store.lenhan.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -8,7 +9,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -20,9 +20,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 
 import com.shoes_store.lenhan.dto.request.brandUpdateDTO;
-import com.shoes_store.lenhan.dto.response.ErrorResponse;
 import com.shoes_store.lenhan.dto.response.brandResponseDTO;
-import com.shoes_store.lenhan.exceptions.brandNotFoundException;
+import com.shoes_store.lenhan.exceptions.ResourceNotFoundException;
 import com.shoes_store.lenhan.model.brand;
 import com.shoes_store.lenhan.services.brandService;
 
@@ -38,33 +37,50 @@ public class brandController {
 
 
 	@GetMapping
-	public List<brand> getAllBrands(){
-		return this.brandservice.getAllBrand();
+	public ResponseEntity<?> getAllBrands(){
+		List<brand> brands = brandservice.getAllBrand();
+		if(brands.isEmpty()) {
+			throw new ResourceNotFoundException("Brand is empty");
+		}
+		return ResponseEntity.ok().body(this.brandservice.getAllBrand());
 	}
 	@GetMapping("/{id}")
-	brand getBrandById(@PathVariable("id") Integer id) {
-		return this.brandservice.getBrandById(id);
+	public ResponseEntity<?> getBrandById(@PathVariable("id") Integer id) {
+		Optional<brand> brand = this.brandservice.getBrandById(id);
+		if(brand.isEmpty()) {
+			throw new ResourceNotFoundException("Brand with id=" + id + "is not exits");
+		}
+		return ResponseEntity.ok().body(this.brandservice.getBrandById(id));
 	}
 	
+//	@PostMapping
+//	@ResponseStatus(HttpStatus.CREATED)
+//	public ResponseEntity<?> createBrand(@Valid @RequestBody brandUpdateDTO dto) {
+//		return ResponseEntity.ok().body(this.brandservice.createBrand(dto));
+//	}
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
-	brandResponseDTO createBrand(@Valid @RequestBody brandUpdateDTO dto) {
+	public brandResponseDTO createBrand(@Valid @RequestBody brandUpdateDTO dto) {
 		return this.brandservice.createBrand(dto);
 	}
 	
 	@PutMapping("/{id}")
-	brandResponseDTO updateBrand(@PathVariable("id") Integer id,@RequestBody brandUpdateDTO dto) {
-		return this.brandservice.updateBrand(id, dto);
+	public ResponseEntity<?> updateBrand(@PathVariable("id") Integer id,@RequestBody brandUpdateDTO dto) {
+		Optional<brand> brand = this.brandservice.getBrandById(id);
+		if(brand.isEmpty()) {
+			throw new ResourceNotFoundException("Brand with id=" + id + "is not exits");
+		}
+		return ResponseEntity.ok().body(this.brandservice.updateBrand(id, dto));
 	}
 	
 	@DeleteMapping("/{id}")
-	public void deleteBrand(@PathVariable Integer id) {
+	public ResponseEntity<?> deleteBrand(@PathVariable("id") Integer id) {
+		Optional<brand> brand = this.brandservice.getBrandById(id);
+		if(brand.isEmpty()) {
+			throw new ResourceNotFoundException("Brand with id=" + id + "is not exits");
+		}
 		this.brandservice.deleteBrand(id);
+		return ResponseEntity.ok().body("Brand with id=" + id + "is  deleted");
 	}
 	
-	@ExceptionHandler({ brandNotFoundException.class })
-	protected ResponseEntity<ErrorResponse> handleBrandNotFoundException() {
-		ErrorResponse error = new ErrorResponse("404", "Brand is Not Found");
-		return new ResponseEntity<ErrorResponse>(error, HttpStatus.NOT_FOUND);
-	}
 }
